+++ src/util/TimetablePrinter.java
@@ -2,56 +2,39 @@ package util;
 
 import config.SchedulingConfig;
 import model.*;
+import ga.Chromosome;
 
 public class TimetablePrinter {
 
-    public static void print(
-            Timetable timetable) {
+    /* Changed parameter type from Timetable to Chromosome */
+    public static void print(Chromosome chromosome) {
 
         String[][] grid = new String[SchedulingConfig.WORKING_DAYS][SchedulingConfig.PERIODS_PER_DAY];
 
-        for (Placement p : timetable.getPlacements()) {
+        for (Placement p : chromosome.getPlacements()) {
 
-            int day = p.getSlot()
-                    .getDay()
-                    .ordinal();
-
-            int start = p.getSlot()
-                    .getPeriod();
-
-            int duration = p.getSession()
-                    .getDuration();
+            int day = p.getSlot().getDay().ordinal();
+            int start = p.getSlot().getPeriod();
+            int duration = p.getSession().getDuration();
 
             for (int i = 0; i < duration; i++) {
-
-                grid[day][start + i] = p.getSession()
-                        .getSubject()
-                        .getCode();
+                grid[day][start + i] = p.getSession().getSubject().getName();
             }
         }
 
         for (int d = 0; d < SchedulingConfig.WORKING_DAYS; d++) {
 
-            System.out.println(
-                    Day.values()[d]);
+            System.out.println(Day.values()[d]);
 
             for (int p = 0; p < SchedulingConfig.PERIODS_PER_DAY; p++) {
 
                 if (grid[d][p] == null)
-                    System.out.printf(
-                            "P%d : FREE%n",
-                            p + 1);
+                    System.out.printf("P%d : FREE%n", p + 1);
                 else
-                    System.out.printf(
-                            "P%d : %s%n",
-                            p + 1,
-                            grid[d][p]);
+                    System.out.printf("P%d : %s%n", p + 1, grid[d][p]);
             }
-
             System.out.println();
 
         }
-
     }
-
 }